fix/: signaux

+++ src/chat_handler.cpp
@@ -17,6 +17,7 @@ int g_file_desc1;
 int g_file_desc2;
 ChatHandler* g_chat_handler = nullptr;
 
+
 ChatHandler::ChatHandler(const string &username1_, const string &username2_, const bool &bot_, const bool &manuel_)
     : user1_name{username1_}, user2_name{username2_}, bot{bot_}, manuel{manuel_} {
     // Tout initialiser
@@ -44,54 +45,43 @@ ChatHandler::ChatHandler(const string &username1_, const string &username2_, con
     }
     // Create Named Pipes (FIFO)
     if (file_desc1 < 0) {
+        std::cout << 'Le probleme c ici file_desc1';
         ExceptionHandler::return_code_check(mkfifo(path_from_user1.c_str(), FIFO_PERMISSION));
     }
     if (file_desc2 < 0) {
+        std::cout << 'Le probleme c ici file_desc2';
         ExceptionHandler::return_code_check(mkfifo(path_from_user2.c_str(), FIFO_PERMISSION));
     }
 }
 void Signal_Handler(const int sig){
     if (sig == SIGINT){
-        if (process > 0){
+        
+        if (g_chat_handler->manuel){
+            signal(SIGINT, SIG_IGN);
+            std::cout << "I am here" << std::endl;
+            g_chat_handler->display_pending_messages();
+            return;
+        }
+        if (process > 0 && !g_chat_handler->manuel){
             std::cout << "You closed the chat. sigint pere" << std::endl;
-            kill(process, SIGTERM);
             close(g_file_desc1);
             close(g_file_desc2);
             unlink(g_path_from_user1.c_str());
             unlink(g_path_from_user2.c_str());
+            kill(0, SIGTERM);  // Terminate both parent and child
             exit(4);
         }
-        else if (process == 0){
-        std::cout << "The other user closed the chat. Signal " << strsignal(sig) << " received." << std::endl;
-        }
+        
     }
-    else if (sig == SIGPIPE) {
-        if (process > 0){
-        printf("Closing the chat.. sigpipe\n");
-        if (g_chat_handler && g_chat_handler->manuel){
-            g_chat_handler->display_pending_messages();
-        }
-        }
-        // Close the fifo channels before removing the files
-        exit(4);
-    /*
-    else if (sig == SIGPIPE){
-        printf("The other user closed the chat. pipesig\n");
-        if (g_chat_handler && g_chat_handler->manuel){
-            g_chat_handler->display_pending_messages();
-        }
-        // Close the fifo channels before removing the files
+        
+    
+    else if(sig == SIGTERM) {
+        std::cout << "Chat closed gracefully (SIGTERM)." << std::endl;
         close(g_file_desc1);
         close(g_file_desc2);
-        std::remove(g_path_from_user1.c_str());
-        std::remove(g_path_from_user2.c_str());
+        unlink(g_path_from_user1.c_str());
+        unlink(g_path_from_user2.c_str());
         exit(4);
-    }
-    else if (sig == SIGTERM){
-        printf("SIGTERM\n");
-        exit(4);
-    }
-    */
     }
 }
 void ChatHandler::access_sending_channel(const string &recipient) {
@@ -121,6 +111,7 @@ void ChatHandler::access_sending_channel(const string &recipient) {
     if (bytes_written < 0) {
         cerr << "Erreur en écriture dans le fichier: " << strerror(errno) << endl << this->error_log << endl;
     } else if (bytes_written == 0) {
+        signal(SIGINT, Signal_Handler);
         string end_user = (path == path_from_user1) ? user1_name : user2_name;
         this->error_log = "Conversation terminée par " + end_user;
         this->exit_code = EXIT_SUCCESS;
@@ -129,7 +120,7 @@ void ChatHandler::access_sending_channel(const string &recipient) {
     exit(this->exit_code);
 }
 void ChatHandler::access_reception_channel(const string &sender) {
-    signal(SIGPIPE, Signal_Handler);
+    //signal(SIGPIPE, Signal_Handler);
     string path = (sender == user2_name) ? path_from_user2 : path_from_user1;
     file_desc2 = open(path.c_str(), O_RDONLY);
     if (file_desc2 < 0) {
@@ -142,21 +133,19 @@ void ChatHandler::access_reception_channel(const string &sender) {
     string ansi_end = bot ? "" : "\x1B[0m";
     do {
         bytes_read = receive_message(received_message);
-        if (bytes_read <= 0) {
+        if (bytes_read < 0) {
             this->error_log = "Erreur de lecture";
             this->exit_code = EXIT_FAILURE;
             cerr << "Erreur en lecture dans le fichier: " << strerror(errno) << endl << this->error_log << endl;
             break;
         }
+        else if(bytes_read == 0) {
+            kill(process, SIGPIPE);
+            break;
+        }
         else if (bytes_read > 0) {
             if (manuel && !bot) {
-                string formatted_message;
-                if (shared_memory_queue->first_message){
-                    formatted_message = "[" + ansi_beginning + sender + ansi_end + "] " + received_message;
-                    shared_memory_queue->first_message = false;
-                } else{
-                    formatted_message = received_message;
-                }
+                string formatted_message = "[" + ansi_beginning + sender + ansi_end + "] " + received_message;
                 add_message_to_shared_memory(formatted_message);
                 pending_bytes += bytes_read;
                 std::cout<< "\a" << std::flush;
@@ -165,7 +154,7 @@ void ChatHandler::access_reception_channel(const string &sender) {
                 fflush(stdout);
             }
         }
-    } while (bytes_read >= 0);
+    } while (bytes_read > 0);
 
     if (bytes_read != 0) {
         this->error_log = "Problème de lecture !";
@@ -194,33 +183,30 @@ int ChatHandler::send_message(char (&message_to_send)[BUFFER_SIZE]){
         perror("sigaction() SIGTERM");
     }
     */
-    //if(FD_ISSET(STDIN_FILENO, &read_fds)){
     char* input = fgets(message_to_send, sizeof(message_to_send), stdin);
     if (input == NULL){
         if (feof(stdin)) {
-            this->error_log = "End of input reached.";
+            this->error_log = "EOF reached. You closed the chat.";
             this->exit_code = EXIT_SUCCESS;
-        } else if (ferror(stdin)) {
-            if (errno == SIGPIPE || errno == SIGINT){
-                std::cout << "Interrupted by signal, the other user closed the chat" << std::endl;
-                this->error_log = "Interrupted by signal, the other user closed the chat.";
-                this->exit_code = EXIT_FAILURE;
-                exit(4);
-            }else{
+            kill(0, SIGTERM);
+            return 0;
+        } 
+        else if (ferror(stdin)) {
             this->error_log = "Error reading input: ";
             this->exit_code = EXIT_FAILURE;
-            }
-        }   
+            return 0;
+        }
+        else{
+            return -1;
+        }
+    }   
+    
+    ssize_t bytes_written = write(file_desc1, message_to_send, strlen(message_to_send));
+    if (bytes_written < 0 && errno == EPIPE) {
+        std::cerr << "Failed to send message: broken pipe." << std::endl;
+        kill(0, SIGTERM);  // Send SIGTERM to close both chats
         return -1;
     }
-    ssize_t bytes_written = write(file_desc1, message_to_send, strlen(message_to_send));
-        if (bytes_written < 0){
-            if (errno == EPIPE) {
-                std::cerr << "Failed to send msg";
-                kill(process, SIGPIPE);
-                return -1;
-            }
-        }
     //this-> exit_code ? bytes_written = -1 : bytes_written;
     return static_cast<int>(bytes_written);
 }
@@ -231,9 +217,11 @@ int ChatHandler::receive_message(char (&received_message)[BUFFER_SIZE]) {
         return -1;
     }
     if (bytes_read == 0) {
-        std::cerr << "The other user has disconnected. \n";
-        kill(process, SIGPIPE);
-        return -1;
+        
+        // I need to know which signal is received so I can treat them differently here.
+
+        kill(0, SIGTERM);
+        //return -1;
         //exit(4); // Fin de la conversation
     }
     received_message[bytes_read] = '\0'; // Null-terminate the string
@@ -243,7 +231,6 @@ void ChatHandler::display_pending_messages() {
     if(shared_memory_queue){
         std::cout.write(shared_memory_queue->messages, shared_memory_queue->total_chars);
         shared_memory_queue->total_chars = 0;
-        shared_memory_queue->first_message = true;
     }
 }
 SharedMemoryQueue* ChatHandler::init_shared_memory_block(){
